requirements.txt
python-telegram-bot==20.7
flask==2.3.3
python-dotenv==1.0.0